Rename External Data Store to Data Link [HZ-2013]
This is a first PR for HZ-2013, which mechanically renames all
references to external data store to external data link.

This PR will make the following PR with changes to the API easier to
review as it will not mix changes done purely for the rename with
changes to how we handle the data store/link, etc..

TODO figure out if we can rename the method in the client protocol.