Remove the hack from test
There was a non-trivial code to create a different, but equivalent condition for
special case, which wasn't very generic and only worked in the present test
cases, only in order for some other, completely unrelated method to notice the
different style of the condition, and use equi-join instead of non-equi-join in
that case. I rather send this one bit of information directly rather than
through coordination of two hacky pieces of code.