Add test for GenericMapStore during shutdown [HZ-1823]
MapStore#destroy is called in 2 cases:
- During Map#destroy - which should destroy the the MapStore as well so
it doesn't leak resources. In this case the mapping should be removed.

- During member shutdown - in this case the mapping should not be
removed because it might be just one member shutting down, not whole
cluster.

Note that the test passes without the new check in destroy method. It is
because when the node is shutting down we can't execute the SQL query to
remove the mapping, however it prevents
`HazelcastInstance is not active!` exception.