Remove incorrect getName override in AbstractParameterizedHazelcastClassRunner (#23132)
The overridden name causes issues in surefire. When the name is not
equal to the name of the class, JUnit creates a different `Description`
here


https://github.com/junit-team/junit4/blob/r4.13.2/src/main/java/org/junit/runners/ParentRunner.java#L395

without a reference to the test class. Surefire then doesn't consider
the class and the categories are incorrectly evaluated.


https://github.com/apache/maven-surefire/blob/surefire-3.0.0-M7/surefire-providers/common-junit48/src/main/java/org/apache/maven/surefire/common/junit48/GroupMatcherCategoryFilter.java#L73

This PR will be accompanied with cleanup in `@Category` annotations.