Change UploadJobMultiPartOperation not to use ForkJoinPool [HZ-2037] (#23600)
UploadJobMultiPartOperation is using ForkJoinPool to process the task,
but depending on the environment (probably the number of cores)
ForkJoinPool is getting stuck. It is better not to derive from
AsyncJobOperation and not to use ForkJoinPool .

The issue was reported in nightly builds because there is a stress test
that has @Category({SlowTest.class}) annotation. This test is getting
stuck and failing the build. The issue link is below

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible