[ALTO] Introduce ClientAltoConfig [ALTO-8] (#23258)
This PR introduces ClientAltoConfig to ClientConfig.

In it, we only have an enable flag for now. Also, all
the configuration elements and schema definitions are documented
as BETA.

This enabled option will have priority over the setSmartRouting
the configuration we have in the ClientNetworkConfig. When Alto config
is enabled, we won't consider the routing set, always try
to use Alto aware mode.

This PR adds the configuration and replaces the places
we use routing mode. Further improvements and tests related
to the Alto-aware client will be added in a further PR.