[ALTO] TPC final config (#23588)
Builders are introduced for AsyncSocket/AsyncServerSocket so that nearly
all fields of the AsyncSocket/AsyncServerSocket can be final so we don't
need to worry about publication problems.

What made this PR more complex is that the Eventloop is now created from
the eventloop thread and not up front. And the eventloop is created
after the affinity is set, so the eventloop can allocate resources
directly on the right core.

This made the creation of the reactor a bit more complex because there
are a bunch of fields like the eventloop and some important dependencies
and I didn't want to make these fields volatile.

So what happens is that the EventloopThread is immediately started and
the Eventloop is created, but the actual processing of the eventloop
starts only when the reactor starts. Using a future, the reactor knows
when the eventloop can be obtained. So all of these fields are now
final.