Moved part of functionality to batch mode only.
While streaming queries return ChangeStreamDocument with fullDocument as the document that was changed, update and sink does not use those wrappers. That makes creating proper mapping difficult.
Would be easier if full scan during update would know it's for the update query (so batch-like), but no such thing exist. In next version custom updates will be possible, so leaving for know as it is