[ALTO] Better Operation thread integration. (#23563)
The old approach created a new type of operation thread by pulling out
an OperationThread interface and introducing a new OperationExecutor
implementation. And compatibility becomes less obvious.

The idea behind the new approach is to subclass the
PartitionOperationThread and override the 'loop' method. So in HZ
classic, there is a loop over the operation queue. In the new approach,
the AltoPartitionOperationThread (which extends the OperationThread)
overrides that loop method and forwards to a Runnable containing the
eventloop. This eventloop loops over a bunch of queues and even sources
(for networking/storage) including the operation queue.

Also, the OperationExecutorImpl is updated so that the appropriate
PartitionOperationThread is created based on TPC being enabled or not.
Apart from that, there is little change.

I believe this should cause the least amount of compatibility problems
with classic OS/EE.