Code against the ClientConnection interface in client related services
We have various client-related services where we code against the Connection
interface, and perform a cast to ClientConnection when it is necessary.

I believe, those casts are ugly, and we are better off coding
against the ClientConnection interface directly so that we won't
need any such casts.

This PR updates various services and tests accordingly.