Fix SessionAwareSemaphoreReleaseAcquiredSessionsOnFailureTest.testTryAcquire_shouldReleaseSessionsOnRuntimeError [HZ-1829] (#23742)
The test is failing because AbstractInvocationFuture, which is a custom
implementation of CompletableFuture needs to spin over a loop to detect
if thread is interrupted. However, if semaphore.tryAcquire() response is
received before the loop, the test fails.
Therefore, we need to make sure that semaphore does not respond. I have
changed the semaphore tryAcquire permit from 1 to 2 to make the
semaphore not respond and allow the AbstractInvocationFuture to spin the
loop to do thread.interrupt check

Jira : https://hazelcast.atlassian.net/browse/HZ-1829
Fixes : https://github.com/hazelcast/hazelcast/issues/20038

There was a similar fix for another test. See 
https://github.com/hazelcast/hazelcast/pull/22995

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible