Decrement owned entry count with proper condition
A record existing in the NearCacheRecordMap does not necessarily
represent an actual value present in the near cache. It can be a
reservation placeholder, placed by a concurrent get operation.
If that is the case, then the owned entry count should not be
decremented because the entry does not yet exist.