[ALTO] Cleanup connection to Alto cores for client [ALTO-8] (#23302)
This PR cleanups the connection logic to the Alto channels
by making sure that the connections are Alto channels
associated with them are in sync. When one closes, the others
do the same.

Also, the closing paths are improved to handle exceptions that
might be thrown.

Another improvement I did is to make the connection to Alto channels
non-blocking and parallel, so that we would achieve lower startup
times on client initialization and reconnects.

Also, some client related tests are added.

Finally, all references to TPC are replaced with Alto to be
consistent.