Fix test_changeJobConfig
The test relied that INSERT doesn't insert any rows, if it fails with
`AccumulationLimitExceededException`. But it might have inserted some rows, and
if it manages to insert them, the query will fail with "Duplicate key" after
restart.

The new test uses a window aggregation with three keys. This fails if max
accumulated records is 2, before emitting anything from the aggregation, so the
INSERT doesn't fail.