Fix test_changeJobConfig again
The test relied on window aggregation, but window aggregation didn't support the
max accumulation limit. It actually failed due to the insert failing, which was
non-deterministic for the same reason as the previous fix.

This PR replaces INSERT with SINK, uses more keys (because they are spread
across multiple processors), and also adds the accumulation check to the
SlidingWindowP.