Remove SO_REUSEPORT from tpcengine tests where not needed.
SO_REUSEPORT should only be needed if you want multiple sockets
listening on the same port. It often is useful for tests that rely
on the same port but the socket isn't fully closed by a previous test.

So we need to monitor spurious test failures when a test fails because
a port is already is in use.