[TPC] Removes SO_REUSEPORT in AltoServerBootstrap. (#23951)
With SO_REUSEPORT you can have multiple sockets listening on the same
port. That isn't needed in this case, we only need 1 socket on that
port. Often we use SO_REUSEPORT to be less picky in unit tests if a port
hasn't been properly closed (yet).

Removing SO_REUSEPORT will prevent problems on platforms not supporting
SO_REUSEPORT. This will resolve a bunch of problems in:

https://github.com/hazelcast/hazelcast/issues/23946