[ALTO] Close connection on congestion. (#23874)
If the response for a client can't be sent due to a congested socket,
the connection is closed.

With classic networking, there is no bound on the write queue for a
socket and hence a single slow client can cause a whole cluster to go
OOME.

With TPC all sockets have buffers with a bound. In this particular case,
we need to deal with the fact that a response to the client can't be
sent when the socket is congested. So when that is detected, the whole
connection (including all the async sockets) is closed to indicate that
there was a problem.