Fix snapshot corrupted after forceful termination [HZ-1230] (#23201)
Jet job snapshot consists in several chunks with processor data (each
with different key which may belong to different partition) written to
`__jet.snapshot.<jobId>.<0 or 1>` map followed by
`SnapshotVerificationRecord` in the same map, and if everything
succeeds, `JobExecutionRecord` is updated to point to the new snapshot.
Jet needs to be sure, that the snapshot chunks are not corrupted/lost in
case of single-node failure (in particular master of partition holding
given chunk of snapshot). Default `IMap` behaviour does not provide such
guarantees because `IMap.put` and other modifying operations return
success in case of backup timeout
(`hazelcast.operation.backup.timeout.millis`) even if synchronous backup
is configured for `IMap`.
If there is risk of losing/corruption of snapshot data, Jet should keep
last known good snapshot (as in case of other errors during snapshot)
and retry later.

If there are problems during 2nd phase of terminal snapshot, we prefer
to restart job so that transactions are promptly committed or rolled
back instead of waiting for restore from the snapshot, which may happen
much later or never.

It can also happen that coordinator fails during 2nd phase of terminal
exported snapshot and will not update Job status. In such case job
should be restarted from the exported snapshot not the automatic one, as
the exported snapshot is the newest.

Fixes #18675

EE PR: https://github.com/hazelcast/hazelcast-enterprise/pull/5692

---------

Co-authored-by: Viliam Durina <viliam@hazelcast.com>
Co-authored-by: Viliam Durina <viliam-durina@users.noreply.github.com>
Co-authored-by: Burak Gök <burak.gok@hazelcast.com>