Use separate messages for TPC authentication We have decided to not add TPC related parameters to the stable authentication codecs, because we are not sure about the final TPC implementation and its requirements. Essentially, these parameters can change in the next releases, but we can't change them in the stable codecs because of the compatibility guarantees we have.
Therefore, we decided to introduce the duplicate experimental client
authentication messages that would contain the TPC-related parameters.
Since the feature is BETA and we won't be providing more than the
same minor version compatibility for it, we can play with these codecs
as we want and remove them while promoting the TPC to stable
and move any additional parameters we have added to the stable codecs.