TPC engine tests less sensitive to dangling ports.
Instead of always binding to port 5000, bind for most test is done
with port=0, meaning that the OS will look up a free port.

This makes the tests less sensitive if a socket is not yet
properly closed by a previous test. In certain cases, it can
happen that the OS keeps a socket around for a bit (e.g.
time_wait state).