AcceptRequest closing.
Functionality has been added for closing the 'accepted' socket.

So the AcceptRequest (and the underlying socket) can be explicitly closed
by calling the close on the AcceptRequest or by throwing an exception.

This is useful for 2 situations:

1) You want to prevent that a socket is 'properly' accepted. Maybe the maximum
number of connection has been reached.

2) There was a problem processing the accept request and we need to ensure that
the socket is properly closed instead of ending up in some zombie state where
the socket is created at the OS level, but not known at the application level.