[TPC] AcceptRequest closing. (#24038)
Functionality has been added for closing the 'accepted' socket from the
AcceptRequest.

So the AcceptRequest (and the underlying socket) can be explicitly
closed by calling the close on the AcceptRequest or by throwing an
exception from the 'acceptConsumer'.

This is useful for 2 situations:

1) You want to prevent a socket is 'properly' accepted. Maybe the
maximum number of sockets has been reached. So you can now call
acceptRequest.close to close the socket for the accept request.

2) There was a problem processing the accept request and we need to
ensure that the socket is properly closed instead of ending up in some
zombie state where the socket is created at the OS level, but not known
at the application level (end hence doesn't get closed). This is
especially a problem with testing where the acceptance of a socket can
be rejected of the reactor has been shut down.