Improve near-cache read-only entry processor tests
There were two main problems with these tests:
- There were lots of repetition and inconsistencies in those repeated
pieces
- Client proxy tests were failing due to invalidation events
coming from the server from time to time.

I have extracted the structure of the tests into a common method
and for the second issue, I put and get entries from the client
so that we won't be dealing with invalidations.