[TPC] Various improvements NioAsyncSocket scheduling (#24096)
1) The NioAsyncSocket is automatically scheduled till the connect has
been completed. This will prevent any write to the socket leading to the
scheduling of the NioAsyncSocket on the reactor. Only when the connect
has been completed, writes to the socket are going to be scheduled.

2) The assert in the NioAsyncSocket.Handler.onWrite isn't valid. If the
SelectionKey is canceled, it will also lead to a call to this method
without the flushThread being set and causes the assert to fail.

3) Various minor cleanups and testing improvements.