Move as many SQL classes to hazelcast-sql module (#24101)
This is a preparation for #23595.

The particular reason for GET_DDL is that there are unparse methods in 
the SQL AST classes. These classes, and the unparsing code itself have
dependencies to Apache Calcite, so they can't be moved to core module,
because Calcite and its dependencies are the main reason for having the
SQL module in the first place.

But because GET_DDL was an expression and all expressions were in the 
root module, it couldn't call back into the SQL module, so we ended up
with two unparse implementations. The new one even didn't handle SQL
escaping correctly. There probably could be a way to hack this around,
but refactoring this first just made sense, and will (hopefully) make
our life easier in the future.

It's also good to move all SQL classes to one module. We've suffered 
from this for a while already, but have been able to work it around
until now.