Rename DataLinkConfig#className to type [HZ-2208] (#24121)
Rename className to type - the reason is that we want to unify the way
the DataLink is created via SQL and via the configuration. In the
configuration, we started with className, but in CREATE DATA LINK we use
type. Also, `type` is more polyglot-friendly.

For 5.3 the following data link types will be supported:
- Jdbc
- Kafka
- Mongo
- Hz - not available via SQL, only remote map journal in Java API

We are allowed to make these changes because we are still in beta for
the data link feature (we already renamed external data store -> data
link).

Client protocol PR
https://github.com/hazelcast/hazelcast-client-protocol/pull/462
EE PR https://github.com/hazelcast/hazelcast-enterprise/pull/5864

Breaking changes (list specific methods/types/messages):
* `className` -> type in DataLinkConfig and the client protocol for
dynamically adding this config