Improve ClientConnectionProcessListenerTest
The test failed with logs indicating that one member we were
expecting to start in port 5701 started in port 5702, which
resulted in failed assertions.

To fix this and some other problems, I have refactored to tests
a bit.

- Now, we don't depend on any specific port and perform our
assertions based on the actual port of the member.
- I have set the cluster connect timeout to 0, so that the
tests we expect the client to fail to connect can complete faster.
- Improved the read-handling code in the `remoteClosesConnection`
test to make sure that the test would pass reliably.
- Removed the unneeded server instance from
`hostResolutionFailure_portFailure` test.
- Used client selectors instead of relying on some implementation
details of the failover client in `clientNotAllowedInCluster` test.