Use TPC tokens for TPC channel authentication in client
This PR adds the necessary protocol changes to use secure-randomly
generated tokens during the authentication of TPC channels from
the client.

Previously, we were using just the client uuid, which was an information
regularly printed in the logs. We would like to be more safe by
using a per-client-connection token to prove the identity claim
of the client when it sends its UUID. (We are aware that this is
not secure as the TLS is not implemented yet for TPC, but it should
be a move in the right direction).

Also, we are now using a normal client message to pass this information
to the server.

The handler on the server side is also improved to handle partial
protocol identifiers.

Also, a document describing the protocol changes we did for
the TPC in 5.3.0 is added, as per request.