Log all exceptions thrown from MapStore/Loader (#23825)
When a MapStore throws an exception it is usually propagated to the
caller and as part of that it is logged to the member log because
operation exceptions are logged.

The call to loadAll() doesn't propagate the exceptions nor logs it.
Added a log on loading future failure.

When a map is destroyed locally (call to `IMap#destroy` on the member),
the exception is propagated to the caller but is not logged. To be
consistent with other operations added a log for such case.