Fix ClientMessageProtectionTest [API-1995] (#24184)
In the failing `testExceededMessageSize`, we are writing a large
authentication message and expecting the connection to close when we try
to read over the socket, which happens all the time.

However, under some orderings, we might even get some IO exceptions
while writing the message. If the server processes the large frame
before the client sends the frames that come after that large frame, the
server might close the connection due to the exceeded message size
limits. And, you can easily get an exception while writing the next
smaller frames.

To solve it, I have moved the expected exception call to before writing
the message line.

closes https://github.com/hazelcast/hazelcast/issues/23467