[TPC] Implement contains method for SelectionKeysSet [HZ-2265] (#24267)
In macos, a single fd can have multiple events after a single poll. Lack
of working `contains()` method creates problems in that case.

To give more detail please look at `updateSelectedKeys` method in
`KQueueSelectorImpl` (there is a link at the bottom):
```
if (selectedKeys.contains(ski)) {
    // first time this file descriptor has been encountered on this
    // update?
    if (me.updateCount != updateCount) {
        if (ski.channel.translateAndSetReadyOps(rOps, ski)) {
            numKeysUpdated++;
            me.updateCount = updateCount;
        }
    } else {
        // ready ops have already been set on this update
        ski.channel.translateAndUpdateReadyOps(rOps, ski);
    }
} else {
    ski.channel.translateAndSetReadyOps(rOps, ski);
    if ((ski.nioReadyOps() & ski.nioInterestOps()) != 0) {
        selectedKeys.add(ski);
        numKeysUpdated++;
        me.updateCount = updateCount;
    }
}
```

Notice that if `contains()` return true and if during this poll key is
already added to the set, `translateAndUpdateReadyOps()` is called, not
`translateAndSetReadyOps()`. The difference is
`translateAndSetReadyOps()`
clears the existing `readyOps` from the key. So in that scenario, we may
lose polled events because previously `contains()` method of the
`SelectorOptimizer` always returned false. So previously, the scenario
was like this:

1. A single file descriptor has 2 events: read(1) and write(4)
2. First time the code enters the above block, `readyOps` will be 1
3. Second time it enters, we put the same object to the set 2nd time
(notice that our set is basically array).
4. But `translateAndSetReadyOps()` is used so `readyOps` will be 4 for
both not 5
5. We lost the read event.

To prevent this we need a working `contains()` implementation.

Since there is a similar failure for Windows, I assume a similar
scenario plays out in Windows's poll implementation.

Note that this issue is present in all java versions I have tested (8,
11, 17, current main branch of openjdk).

https://github.com/openjdk/jdk8u-dev/blob/f1c1f969a77c9841cf87b2434fa90dcf4002c6be/jdk/src/macosx/classes/sun/nio/ch/KQueueSelectorImpl.java#L130

Fixes MacOS and Windows related failures of:

- https://github.com/hazelcast/hazelcast/issues/23867
- https://github.com/hazelcast/hazelcast/issues/23823
- https://github.com/hazelcast/hazelcast/issues/24076
- https://github.com/hazelcast/hazelcast/issues/23956