Tighten up DynamicMapConfigTest and add more verbose failure messages
It was possible for this test to fail during the eviction testing phase,
but not provide enough information as to why it failed. It was not clear
whether it failed to evict the *correct* entry, or whether eviction
failed entirely.

I have tightened up this test to provide more detailed failure messages
if the eviction phase does not succeed, such as which key was evicted
if it was not the expected key.

Closes https://github.com/hazelcast/hazelcast/issues/23762