Create pool in JdbcDataConnection lazily [HZ-2317]
When the pool creation fails and throws exception we get out of synch in
InternalDataLinkService and in SQL catalog of data connections.

This change creates the pool lazily using ConcurrentMemoizingSupplier.
The exception from pool creation will be thrown to the caller of
`JdbcDataConnection#getConnection`, e.g. when `CREATE MAPPING ...` is
excecuted.

Fixes #2431