Support catalog and schema in listResources [HZ-2325] (#24339)
Fixes a leftover todo and a bug reported by MC for MySQL - MC-2048.

The resource name consists of multiple parts. These are supposed to be
used for the external name when creating the mapping.
What those parts are is specific to the 3rd party system. E.g. each
relational database handles catalog/schema/table name differently.

The output of `SHOW RESOURCES FOR ...` reflects this change and returns
quoted compound identifier, similar to `externalName` in
`information_schema.mappings`.