Ignore ConfigMismatchOp shutdown requests if already in a cluster
As demonstrated in `TcpIpJoinTest#test_whenIncompatiblePartitionGroups`
test failures, due to the nature of operation retry mechanics, it is
possible for a younger node to receive a delayed `WhoIsMasterOp` from an
older node which has already established itself as master. This can lead
to the younger node responding with a `ConfigMismatchOp` if there is a
configuration mismatch, which results in the older node being shutdown
after processing the `ConfigMismatchOp`.

Since we can be sure that a node's configuration has been validated if
it is already part of a cluster, we can ignore shutdown requests from a
`ConfigMismatchOp` if our node returns true for `ClusterManager#isJoined`.
This ensures that an already established node which is part of a cluster
can no longer be shutdown by the race condition which is possible when
`WhoIsMasterOp` operations are submitted after a retry delay.

Fixes https://github.com/hazelcast/hazelcast/issues/23384