Simplify ClientRegressionWithMockNetworkTest (#24292)
In an earlier PR, we have added a latch to wait for a thread we started
in a test to finish. And, we have added a check to make sure that there
are no leftover instances.

For the first one, I believe waiting for the thread termination is just
enough.

For the second one, the added check does not verify anything as I wrote
in the original issue description. It feels weird to have it right after
terminateAll call. So, I would like to remove it as well.

I will backport this simplified version to other 5.x branches.