Implement client-side support for TPC enabled clusters on Viridian (#24378)
This PR enables the client to connect to clusters running on Viridian
with TPC enabled.

The new implementation relies on the agreed changes to the coordinator
endpoint response, which now contains the private to public TPC port
mappings if the TPC is enabled on the cluster.

The client now stores the private to public address mappings of the TPC
ports and does translation before connecting to TPC addresses. Also,
since the TPC does not yet support TLS, it always uses plain text
channels for TPC connections.