Replace Thread#sleep with CountDownLatch in SerialOperationInvokerTest
This provides no functional difference, but is more appropriate
for consistency within the tests, and a generally cleaner approach.

This commit also removes AtomicInteger resetting in the `afterClass()`
function since it no longer serves purpose as the introduction of
latches prevents a clean loop without them being reset as well.