Introduce CountdownLatch in SerialOperationInvokerTest restarting tests  [HZ-2341] (#24392)
For context: the aim of 2 tests in this class are to ensure that a set
of serial operations are executed and successfully restarted when the
cluster topology changes; there is one test for removing a member and
one for adding a new member. Both tests should record a complete ordered
list of member UUIDs as they exist before the topology change, and then
record the new list of member UUIDs in the same collection after the
topology change.

Under certain conditions, it was possible for the topology change
triggered by adding/removing a cluster member to occur *before* the
initial recording operation was finished due to race conditions. The
method used in this test for serial invocation on the cluster
(`InvocationUtil#invokeOnStableClusterSerial`) normally expects
idempotent operations, but due to the nature of the test being performed
(where we want the originally recorded values to exist in addition to
the restarted operation values), our test slightly violates this
agreement.

We can resolve this conflict by introducing a simple CountdownLatch in
each of the tests that allows us to wait for the initial recording
operations to complete before changing the topology of the cluster and
causing the operations to be restarted. This removes the race condition
and allows us to produce concrete results for both tests.

Fixes https://github.com/hazelcast/hazelcast/issues/24007