Try to load the class for the type name of the Compact GenericRecord once
Compact serialization supports zero-configuration. In that case,
it uses the fully qualified class name as its type name.

While reading Compact classes, when we can't find a registration
for the given type name, there are three possible cases:

- It was serialized with zero-config
- It was serialized as GenericRecord
- It was serialized with an explicit serializer, but there
is no configuration for it on the reader side.

When it is serialized with zero-config and we have the class in
our classpath, the read operation should return an instance
of that class.

To do that, we were trying to load the class associated with
the type name of the Compact object. If that is successful,
we assume that it was serialized with zero-config. If not,
we know that the object must be read as GenericRecord.

There is nothing wrong with this approach, but we were not
caching the result of this class loading operation. That means,
while reading GenericRecords, we were constantly trying to
load a class with the given typeName, which might hurt
the performance.

To fix it, we now have a special registration that can be used
to identify typeNames that must read as GenericRecords. And,
when the class loading operation fails for a given type name,
we cache this special value to signal for later operations
that we should read the Compact object as GenericRecord.

I have also removed the leftover try-catch block that we
had while creating a registration for our reflective serializers.

Also, I have fixed some tests to use GenericRecords, while
we assume the server does not have classes in its classpath.